+++ src/app/app.module.ts
@@ -394,6 +394,7 @@ import { KnowledgeBaseTableComponent } from './knowledge-bases/modals/knowledge-
 import { ModalErrorComponent } from './knowledge-bases/modals/modal-error/modal-error.component';
 import { KnowledgeBasesPreviousComponent } from './knowledge-bases-previous/knowledge-bases-previous.component';
 import { ModalUrlsKnowledgeBaseComponent } from './knowledge-bases/modals/modal-urls-knowledge-base/modal-urls-knowledge-base.component';
+import { UserModalComponent } from './users/user-modal/user-modal.component';
 
 
 
@@ -698,7 +699,8 @@ const appInitializerFn = (appConfig: AppConfigService, brandService: BrandServic
     KnowledgeBaseTableComponent,
     ModalErrorComponent,
     KnowledgeBasesPreviousComponent,
-    ModalUrlsKnowledgeBaseComponent
+    ModalUrlsKnowledgeBaseComponent,
+    UserModalComponent
   ],
   imports: [
     TooltipModule.forRoot(CutomTooltipOptions as TooltipOptions),